Move FeatureExtractor to its own feature_extractor module
